requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromedriver-autoinstaller==0.6.4
click==8.2.1
colorama==0.4.6
configargparse==1.7.1
exceptiongroup==1.3.0
flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
locust==2.37.11
locust-cloud==1.24.2
MarkupSafe==3.0.2
msgpack==1.1.1
nest-asyncio==1.6.0
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydentic==0.0.1.dev3
pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
pytest-check==2.5.3
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
python-stdnum==2.1
python-telegram-bot==22.1
pywin32==310
pyzbar==0.1.9
pyzmq==27.0.0
requests==2.32.4
selenium==4.34.2
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
super-puer-frame==0.7.0
super-puper-frame==0.8.0
termcolor==3.1.0
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
utils-hj3415==3.2.4
webdriver-hj3415==2.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
werkzeug==3.1.3
wsproto==1.2.0
zope-event==5.1
zope.interface==7.2
